backend/requirements.txt
fastapi[all]==0.110.0
uvicorn==0.29.0
openai==1.14.3
langchain==0.1.13
langchain-openai==0.1.1
pydantic==2.6.4
python-dotenv==1.0.1
python-multipart==0.0.9
